Move browser helpers into platform
