+++ web/src/routes/(app)/preferences/WorkAsRemoteSigner.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { _ } from 'svelte-i18n';
 	import { remoteSigner } from '$lib/RemoteSigner';
-	import { copy as copyToClipboard } from '$lib/Clipboard';
+	import { copy as copyToClipboard } from '$lib/platform/browser/clipboard';
 	import { IconCheck } from '@tabler/icons-svelte-runes';
 	import { page } from '$app/state';
